fix bug's

+++ Assets/Scripts/Iinstance.cs
@@ -10,7 +10,7 @@ public class Iinstance : MonoBehaviour
     public static Iinstance instance;
     public int Coins;
     public int FreeSpins;
-    public Dictionary<string, int> TypesOfTask = new(AmountOfTask);
+    public Dictionary<string, float> TypesOfTask = new(AmountOfTask);
     public string[] ArrayOfNameTask = new string[AmountOfTask];
     private void Awake()
     {
